Use int instead of size_t for string length.  All other flesh routines also use int for lengths and not size_t.
git-svn-id: https://svn.cactuscode.org/flesh/trunk@3950 17b73243-c579-4c4c-a9d2-2d5706c11dac

+++ src/include/cctk_Misc.h
@@ -22,7 +22,7 @@ int Util_SplitString(char **before, char **after, const char *string, const char
 int CCTK_Equals(const char *string1, const char *string2);
 int CCTK_FortranString (const char *c_string,
                         char *fortran_string,
-                        size_t fortran_length);
+                        int fortran_length);
 
 char *Util_NullTerminateString(const char *, unsigned int);
